Fix issue that was preventing to use service-dns when installed via helm chart
As the property service.name is defined by the default in the Hazelcast chart and it cannot be used in conjunction with service-dns, it's needed to use an empty value for service.name, what is not allowed in the current version. This same fix was already done in the version 2.x, but it still was not fixed in the version 1.5.x, which prevents to use service-dns to connect with the version 3.x of Hazelcast.